Split parser into hpp and cpp
