+++ include/Parser.h
@@ -1,10 +1,15 @@
 #pragma once
 
-#include <functional>
 #include <string>
-#include "Model.h"
 
 namespace Parser {
+
+struct Node {
+    virtual long double calc() const noexcept = 0;
+};
+
+
+Node *parse_expression(const std::string &x_expr);
 
 auto parse_expressions(const std::string &x_expr, const std::string &y_expr, const std::string &z_expr);
